// src/app/modules/dashboard/util/enums/routes/calendar.ts
export enum CalendarRoutes {
  BASE = '/calendar',
  GET_ALL_BY_PERIOD = '/getAllByPeriod',
  GET_ALL_TYPES = '/getAllTypes',
  GET_PLANIFICATION_CALENDAR_BY_PERIOD = '/getPlanificationCalendarByPeriod',
  FINISH_PLANIFICATION_CALENDAR = '/finishPlanificationCalendarByPeriod',
  CREATE_CALENDAR = '/createNewCalendar'
}